llm-service/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-huggingface
langchain-chroma
langchain-ollama
chromadb
sentence-transformers
python-dotenv
psycopg2-binary
